Rename error_handlers -> globalErrorHandlers. File & imports
