+++ packages/jest-circus/src/eventHandler.js
@@ -20,7 +20,7 @@ import {
 import {
   injectGlobalErrorHandlers,
   restoreGlobalErrorHandlers,
-} from './error_handlers';
+} from './globalErrorHandlers';
 
 // To pass this value from Runtime object to state we need to use global[sym]
 const TEST_TIMEOUT_SYMBOL = Symbol.for('TEST_TIMEOUT_SYMBOL');
